[misc] Move ThreadResourceResolverProvider into its own module, since it is needed in an early-accessible bundle and not all of cards-utils can start that early
