+++ heracles-resources/backend/pom.xml
@@ -63,6 +63,11 @@
     </dependency>
     <dependency>
       <groupId>${project.groupId}</groupId>
+      <artifactId>cards-resolver-provider</artifactId>
+      <version>${project.version}</version>
+    </dependency>
+    <dependency>
+      <groupId>${project.groupId}</groupId>
       <artifactId>cards-utils</artifactId>
       <version>${project.version}</version>
     </dependency>
