+++ heracles-resources/backend/src/main/java/io/uhndata/cards/heracles/internal/PauseResumeFormEditor.java
@@ -41,8 +41,8 @@ import org.slf4j.LoggerFactory;
 
 import io.uhndata.cards.forms.api.FormUtils;
 import io.uhndata.cards.forms.api.QuestionnaireUtils;
+import io.uhndata.cards.resolverProvider.ThreadResourceResolverProvider;
 import io.uhndata.cards.subjects.api.SubjectUtils;
-import io.uhndata.cards.utils.ThreadResourceResolverProvider;
 
 /**
  * An {@link Editor} that fills out any reference answers for a new form.
