+++ heracles-resources/backend/src/main/java/io/uhndata/cards/heracles/internal/export/ExportEndpoint.java
@@ -34,7 +34,7 @@ import org.apache.sling.servlets.annotations.SlingServletResourceTypes;
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
 
-import io.uhndata.cards.utils.ThreadResourceResolverProvider;
+import io.uhndata.cards.resolverProvider.ThreadResourceResolverProvider;
 
 @Component(service = { Servlet.class })
 @SlingServletResourceTypes(
