+++ modules/data-model/forms/impl/pom.xml
@@ -108,6 +108,11 @@
     </dependency>
     <dependency>
       <groupId>${project.groupId}</groupId>
+      <artifactId>cards-resolver-provider</artifactId>
+      <version>${project.version}</version>
+    </dependency>
+    <dependency>
+      <groupId>${project.groupId}</groupId>
       <artifactId>cards-dataentry</artifactId>
       <version>${project.version}</version>
     </dependency>
