+++ modules/data-model/forms/impl/src/main/java/io/uhndata/cards/forms/internal/FormRelatedSubjectsEditorProvider.java
@@ -29,7 +29,7 @@ import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
 import org.osgi.service.component.annotations.ServiceScope;
 
-import io.uhndata.cards.utils.ThreadResourceResolverProvider;
+import io.uhndata.cards.resolverProvider.ThreadResourceResolverProvider;
 
 /**
  * A {@link EditorProvider} returning {@link FormRelatedSubjectsEditor}.
