+++ modules/data-model/subjects/impl/pom.xml
@@ -94,6 +94,11 @@
     </dependency>
     <dependency>
       <groupId>${project.groupId}</groupId>
+      <artifactId>cards-resolver-provider</artifactId>
+      <version>${project.version}</version>
+    </dependency>
+    <dependency>
+      <groupId>${project.groupId}</groupId>
       <artifactId>cards-dataentry</artifactId>
       <version>${project.version}</version>
     </dependency>
