+++ modules/patient-portal/pom.xml
@@ -129,6 +129,11 @@
       <artifactId>javax.servlet-api</artifactId>
     </dependency>
     <dependency>
+      <groupId>${project.groupId}</groupId>
+      <artifactId>cards-resolver-provider</artifactId>
+      <version>${project.version}</version>
+    </dependency>
+    <dependency>
       <groupId>io.uhndata.cards</groupId>
       <artifactId>cards-dataentry</artifactId>
       <version>${project.version}</version>
