+++ modules/patient-portal/src/main/java/io/uhndata/cards/patients/emailnotifications/GeneralNotificationsTask.java
@@ -38,7 +38,7 @@ import io.uhndata.cards.emailnotifications.EmailTemplate;
 import io.uhndata.cards.forms.api.FormUtils;
 import io.uhndata.cards.metrics.Metrics;
 import io.uhndata.cards.patients.api.PatientAccessConfiguration;
-import io.uhndata.cards.utils.ThreadResourceResolverProvider;
+import io.uhndata.cards.resolverProvider.ThreadResourceResolverProvider;
 
 public class GeneralNotificationsTask extends AbstractEmailNotification implements Runnable
 {
