+++ modules/permissions/src/main/java/io/uhndata/cards/permissions/internal/SessionSubjectRestrictionFactory.java
@@ -27,7 +27,7 @@ import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
 
 import io.uhndata.cards.permissions.spi.RestrictionFactory;
-import io.uhndata.cards.utils.ThreadResourceResolverProvider;
+import io.uhndata.cards.resolverProvider.ThreadResourceResolverProvider;
 
 /**
  * Factory for {@link SessionSubjectRestrictionPattern}.
