+++ modules/pom.xml
@@ -32,6 +32,7 @@
 
   <modules>
     <module>startup-customization</module>
+    <module>resolver-provider</module>
     <module>form-completion-status</module>
     <module>commons</module>
     <module>ui-extension</module>
