+++ modules/resolver-provider/src/main/java/io/uhndata/cards/resolverProvider/ThreadResourceResolverProvider.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.uhndata.cards.utils;
+package io.uhndata.cards.resolverProvider;
 
 import org.apache.sling.api.resource.ResourceResolver;
