+++ modules/scheduled-csv-export/src/main/java/io/uhndata/cards/scheduledcsvexport/ExportEndpoint.java
@@ -35,7 +35,7 @@ import org.apache.sling.servlets.annotations.SlingServletResourceTypes;
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
 
-import io.uhndata.cards.utils.ThreadResourceResolverProvider;
+import io.uhndata.cards.resolverProvider.ThreadResourceResolverProvider;
 
 @Component(service = { Servlet.class })
 @SlingServletResourceTypes(
