+++ modules/scheduled-csv-export/src/main/java/io/uhndata/cards/scheduledcsvexport/ScheduledExport.java
@@ -32,7 +32,7 @@ import org.osgi.service.component.annotations.ReferencePolicyOption;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import io.uhndata.cards.utils.ThreadResourceResolverProvider;
+import io.uhndata.cards.resolverProvider.ThreadResourceResolverProvider;
 
 @Component(immediate = true)
 public class ScheduledExport
