+++ modules/vocabularies/pom.xml
@@ -73,7 +73,7 @@
   <dependencies>
     <dependency>
       <groupId>${project.groupId}</groupId>
-      <artifactId>cards-utils</artifactId>
+      <artifactId>cards-resolver-provider</artifactId>
       <version>${project.version}</version>
     </dependency>
     <dependency>
