+++ modules/vocabularies/src/main/java/io/uhndata/cards/vocabularies/internal/DefaultBioPortalApiKeyManager.java
@@ -27,7 +27,7 @@ import org.osgi.service.component.annotations.Reference;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import io.uhndata.cards.utils.ThreadResourceResolverProvider;
+import io.uhndata.cards.resolverProvider.ThreadResourceResolverProvider;
 import io.uhndata.cards.vocabularies.BioPortalApiKeyManager;
 
 /**
