+++ proms-resources/backend/pom.xml
@@ -113,6 +113,11 @@
       <artifactId>javax.servlet-api</artifactId>
     </dependency>
     <dependency>
+      <groupId>${project.groupId}</groupId>
+      <artifactId>cards-resolver-provider</artifactId>
+      <version>${project.version}</version>
+    </dependency>
+    <dependency>
       <groupId>io.uhndata.cards</groupId>
       <artifactId>cards-dataentry</artifactId>
       <version>${project.version}</version>
